refactor: encrypt utill 함수 분리 - #1
