+++ src/pages/LogIn/LogInForm.tsx
@@ -1,84 +1,32 @@
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
-import { useLocation } from 'react-router-dom';
-import { JSEncrypt } from 'jsencrypt';
 import { getTerms, logIn, requestKey } from '../../api/LogIn.ts';
+import getAuth from '../../utils/getAuth.ts';
+import { setEncryptData } from '../../utils/encryptionUtils.ts';
+
 import LogInInput from './LogInInput.tsx';
 import LogInButton from './LogInButton.tsx';
 import CheckBox from './CheckBox.tsx';
-import {
-  AuthFormValues,
-  EncryptKeyInfo,
-  LogInFormDataValues,
-  Terms,
-} from '../../types';
-
-const encrypt = new JSEncrypt();
-const encryptFunc = (value: string, key: string) => {
-  if (!key) {
-    throw new Error('encrypt key가 유효하지 않습니다.');
-  }
-  encrypt.setPublicKey(key);
-  return encrypt.encrypt(value);
-};
-
-const setEncryptData = (
-  formData: AuthFormValues,
-  encryptInfo: EncryptKeyInfo,
-  path: string,
-  terms?: Map<string, boolean>
-): LogInFormDataValues => {
-  const { rsaPublicKey, credentialKey } = encryptInfo;
-
-  const loginFormData: LogInFormDataValues = {
-    encryptedPassword: '',
-    key: '',
-  };
-
-  if (path === '/login') {
-    loginFormData.encryptedStudentId =
-      encryptFunc(formData.id, rsaPublicKey) || '';
-    loginFormData.terms = terms;
-  } else if (path === '/admin/login') {
-    loginFormData.encryptedLoginId =
-      encryptFunc(formData.id, rsaPublicKey) || '';
-  }
-  loginFormData.encryptedPassword =
-    encryptFunc(formData.password, rsaPublicKey) || '';
-  loginFormData.key = credentialKey;
-
-  return loginFormData;
-};
+import { AuthFormValues, Terms } from '../../types';
 
 const LogInForm = () => {
   const [termsList, setTermsList] = useState<Terms[]>([]);
   const { register, handleSubmit } = useForm<AuthFormValues>();
-  const location = useLocation();
-
-  const getAuth = () => {
-    if (location.pathname === '/login') {
-      return 'user';
-    }
-    if (location.pathname === '/admin/login') {
-      return 'admin';
-    }
-    throw new Error('로그인 외 경로');
-  };
+  const auth = getAuth();
 
   const onSubmit = handleSubmit(async (formData) => {
     try {
       const encryptInfo = await requestKey();
       const terms = new Map<string, boolean>(
-        termsList.map((term) => [term.id, formData?.terms?.[term.id]] ?? false)
+        termsList.map((term) => [term.id, formData?.terms?.[term.id] ?? false])
       );
-
       const encryptLogInData = setEncryptData(
         formData,
         encryptInfo,
-        location.pathname,
+        auth,
         terms
       );
-      await logIn(encryptLogInData, getAuth(), encryptInfo.rsaKeyStrategy);
+      await logIn(encryptLogInData, auth, encryptInfo.rsaKeyStrategy);
     } catch (e) {
       if (e instanceof Error) {
         console.error(e.message);
@@ -111,7 +59,7 @@ const LogInForm = () => {
         placeholder="비밀번호를 입력해주세요"
         register={register}
       />
-      {location.pathname === '/login' &&
+      {auth === 'user' &&
         termsList.map((t) => (
           <div key={t.id}>
             <p>{t.title}</p>
